Refactor DI for Key Vault service
